requirements.txt
apache-airflow
pandas
rapidfuzz
sqlalchemy
psycopg2-binary
python-dotenv
